+++ internal/grpcwrapper/rawtopic/describe_topic.go
@@ -18,6 +18,7 @@ type DescribeTopicRequest struct {
 	OperationParams rawydb.OperationParams
 	Path            string
 	IncludeStats    bool
+	IncludeLocation bool
 }
 
 func (req *DescribeTopicRequest) ToProto() *Ydb_Topic.DescribeTopicRequest {
@@ -25,6 +26,7 @@ func (req *DescribeTopicRequest) ToProto() *Ydb_Topic.DescribeTopicRequest {
 		OperationParams: req.OperationParams.ToProto(),
 		Path:            req.Path,
 		IncludeStats:    req.IncludeStats,
+		IncludeLocation: req.IncludeLocation,
 	}
 }
 
@@ -99,6 +101,7 @@ type PartitionInfo struct {
 	ChildPartitionIDs  []int64
 	ParentPartitionIDs []int64
 	PartitionStats     PartitionStats
+	PartitionLocation  PartitionLocation
 	FromBound          []byte
 	ToBound            []byte
 }
@@ -116,5 +119,23 @@ func (pi *PartitionInfo) FromProto(proto *Ydb_Topic.DescribeTopicResult_Partitio
 		pi.ToBound = keyRange.GetToBound()
 	}
 
+	pi.PartitionLocation.FromProto(proto.GetPartitionLocation())
+
 	return pi.PartitionStats.FromProto(proto.GetPartitionStats())
+}
+
+// PartitionLocation describes which node currently hosts a partition.
+// It is populated only when DescribeTopicRequest.IncludeLocation is set to true.
+type PartitionLocation struct {
+	NodeID     int32
+	Generation int64
+}
+
+// FromProto fills location from proto. Accepts nil (location info not requested).
+func (pl *PartitionLocation) FromProto(proto *Ydb_Topic.PartitionLocation) {
+	if proto == nil {
+		return
+	}
+	pl.NodeID = proto.GetNodeId()
+	pl.Generation = proto.GetGeneration()
 }