+++ internal/topic/topicmultiwriter/config.go
@@ -8,6 +8,7 @@ type MultiWriterConfig struct {
 	WriterIdleTimeout time.Duration
 	ProducerIDPrefix  string
 	PartitionChooser  PartitionChooser
+	DirectWrite       bool
 
 	writersFactory writersFactory
 }