+++ internal/topic/topicmultiwriter/partition_writer_pool.go
@@ -100,6 +100,10 @@ func (p *partitionWriterPool) createDirectWriter(partitionID int64) (writer, err
 		}
 	)
 
+	if p.cfg.DirectWrite {
+		opts = append(opts, topicwriterinternal.WithDirectWrite(true))
+	}
+
 	writerCfg.MultiMode = true
 	for _, opt := range opts {
 		opt(&writerCfg)